+++ src/features/pre-opinion/components/PreOpinionWriteHeader.tsx
@@ -1,6 +1,5 @@
-import { useEffect, useRef, useState } from 'react'
-
 import type { PreOpinionBook } from '@/features/pre-opinion/preOpinion.types'
+import { useScrollShadow } from '@/shared/hooks'
 import { cn } from '@/shared/lib/utils'
 import { Button } from '@/shared/ui'
 
@@ -42,62 +41,43 @@ const PreOpinionWriteHeader = ({
   isSubmitting,
   isReviewValid = false,
 }: PreOpinionWriteHeaderProps) => {
-  const sentinelRef = useRef<HTMLDivElement>(null)
-  const [isStuck, setIsStuck] = useState(false)
-
-  useEffect(() => {
-    const sentinel = sentinelRef.current
-    if (!sentinel) return
-
-    const observer = new IntersectionObserver(
-      ([entry]) => {
-        setIsStuck(!entry.isIntersecting)
-      },
-      { threshold: 0 }
-    )
-
-    observer.observe(sentinel)
-    return () => observer.disconnect()
-  }, [])
+  const isScrolled = useScrollShadow()
 
   return (
-    <>
-      <div ref={sentinelRef} className="h-0" aria-hidden />
-      <div
-        className={cn(
-          'sticky top-[123px] z-30 bg-white w-screen ml-[calc(-50vw+50%)] transition-shadow duration-200',
-          isStuck && '[box-shadow:0_6px_6px_-4px_rgba(17,17,17,0.08)]'
-        )}
-      >
-        <div className="mx-auto max-w-layout-max px-layout-padding h-[65px] pb-tiny">
-          <div className="flex justify-between items-center">
-            <div className="flex flex-col gap-xtiny">
-              <h3 className="typo-heading3 text-black">사전 의견 작성하기</h3>
-              <p className="text-grey-600">
-                {book.title} · {book.author}
-              </p>
-            </div>
-            <div className="flex items-center">
-              {updatedAt && (
-                <p className="typo-body6 text-grey-600 mr-large">{formatUpdatedAt(updatedAt)}</p>
-              )}
-              <Button
-                className="mr-xsmall"
-                variant={'secondary'}
-                outline
-                onClick={onSave}
-                disabled={isSaving || !isReviewValid}
-              >
-                {isSaving ? '저장 중...' : '저장하기'}
-              </Button>
-              <Button onClick={onSubmit} disabled={isSubmitting || isSaving || !isReviewValid}>
-                {isSubmitting ? '공유 중...' : '공유하기'}
-              </Button>
-            </div>
+    <div
+      className={cn(
+        'sticky top-[123px] z-30 bg-white transition-shadow duration-200',
+        isScrolled && 'shadow-drop-bottom'
+      )}
+    >
+      <div className="mx-auto max-w-layout-max px-layout-padding h-[65px] pb-tiny">
+        <div className="flex justify-between items-center">
+          <div className="flex flex-col gap-xtiny">
+            <h3 className="typo-heading3 text-black">사전 의견 작성하기</h3>
+            <p className="text-grey-600">
+              {book.title} · {book.author}
+            </p>
+          </div>
+          <div className="flex items-center">
+            {updatedAt && (
+              <p className="typo-body6 text-grey-600 mr-large">{formatUpdatedAt(updatedAt)}</p>
+            )}
+            <Button
+              className="mr-xsmall"
+              variant={'secondary'}
+              outline
+              onClick={onSave}
+              disabled={isSaving || !isReviewValid}
+            >
+              {isSaving ? '저장 중...' : '저장하기'}
+            </Button>
+            <Button onClick={onSubmit} disabled={isSubmitting || isSaving || !isReviewValid}>
+              {isSubmitting ? '공유 중...' : '공유하기'}
+            </Button>
           </div>
         </div>
       </div>
-    </>
+    </div>
   )
 }
