+++ src/features/pre-opinion/hooks/index.ts
@@ -1,4 +1,6 @@
 export * from './preOpinionQueryKeys'
+export * from './useDeleteMyPreOpinionAnswer'
 export * from './usePreOpinion'
+export * from './usePreOpinionAnswers'
 export * from './useSavePreOpinion'
 export * from './useSubmitPreOpinion'
