+++ src/features/pre-opinion/hooks/usePreOpinionAnswers.ts
@@ -7,8 +7,8 @@ import { useQuery } from '@tanstack/react-query'
 
 import type { ApiError } from '@/api'
 
-import { getPreOpinionAnswers } from '../preOpinions.api'
-import type { GetPreOpinionAnswersParams, PreOpinionAnswersData } from '../preOpinions.types'
+import { getPreOpinionAnswers } from '../preOpinion.api'
+import type { GetPreOpinionAnswersParams, PreOpinionAnswersData } from '../preOpinion.types'
 import { preOpinionQueryKeys } from './preOpinionQueryKeys'
 
 /**
@@ -18,9 +18,7 @@ import { preOpinionQueryKeys } from './preOpinionQueryKeys'
  * TanStack Query를 사용하여 약속의 사전 의견 목록을 조회합니다.
  * 멤버별 책 평가, 주제별 의견 등을 포함합니다.
  *
- * @param params - 조회 파라미터
- * @param params.gatheringId - 모임 식별자
- * @param params.meetingId - 약속 식별자
+ * @param params - 모임 ID와 약속 ID
  *
  * @returns TanStack Query 결과 객체
  */
