--- src/features/preOpinions/components/index.ts
@@ -1,2 +1,0 @@
-export * from './PreOpinionDetail'
-export * from './PreOpinionMemberList'
