--- src/features/preOpinions/index.ts
@@ -1,13 +1,0 @@
-// Components
-export * from './components'
-
-// Hooks
-export * from './hooks'
-
-// API
-export * from './preOpinions.api'
-export * from './preOpinions.endpoints'
-export * from './preOpinions.mock'
-
-// Types
-export * from './preOpinions.types'
